+++ addons/hr/static/src/new/messaging.js
@@ -4,12 +4,16 @@ import { Messaging } from "@mail/new/core/messaging";
 
 import { patch } from "web.utils";
 import { _t } from "@web/core/l10n/translation";
+import { ThreadService } from "@mail/new/thread/thread_service";
 
 patch(Messaging.prototype, "hr", {
     setup(...args) {
         this._super(...args);
         this.state.employees = {};
     },
+});
+
+patch(ThreadService.prototype, "hr", {
     async getChat(person) {
         const { employeeId } = person;
         const _super = this._super.bind(this);
