+++ addons/mail/static/src/new/thread/thread.xml
@@ -60,7 +60,7 @@
 </t>
 
 <t t-name="mail.thread.load_more" owl="1">
-    <button class="btn btn-link" t-on-click="() => messaging.fetchThreadMessagesMore(props.thread)" t-ref="load-more">Load More</button>
+    <button class="btn btn-link" t-on-click="() => threadService.fetchMoreMessages(props.thread)" t-ref="load-more">Load More</button>
 </t>
 
 <t t-name="mail.notification_message" owl="1">
