+++ addons/mail/static/tests/helpers/webclient_setup.js
@@ -30,6 +30,7 @@ import { rtcService } from "@mail/new/rtc/rtc_service";
 import { suggestionService } from "@mail/new/suggestion/suggestion_service";
 import { stateService } from "@mail/new/core/state_service";
 import { chatWindowService } from "@mail/new/chat/chat_window_service";
+import { threadService } from "@mail/new/thread/thread_service";
 
 const ROUTES_TO_IGNORE = [
     "/web/webclient/load_menus",
@@ -100,6 +101,7 @@ function setupMessagingServiceRegistries({ loadingBaseDelayDuration = 0, messagi
         effect: effectService,
         "mail.suggestion": suggestionService,
         "mail.state": stateService,
+        "mail.thread": threadService,
         "mail.chat_window": chatWindowService,
         "mail.messaging": newMessagingService,
         "mail.rtc": rtcService,
